+++ src/main/java/gift/product/application/ProductService.java
@@ -5,8 +5,6 @@ import gift.global.error.ErrorCode;
 import gift.product.dao.CategoryRepository;
 import gift.product.dao.OptionRepository;
 import gift.product.dao.ProductRepository;
-import gift.product.dto.OptionRequest;
-import gift.product.dto.OptionResponse;
 import gift.product.dto.ProductRequest;
 import gift.product.dto.ProductResponse;
 import gift.product.entity.Category;
@@ -17,8 +15,6 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.Set;
 
 @Service
 public class ProductService {
@@ -46,10 +42,6 @@ public class ProductService {
                 .orElseThrow(() -> new CustomException(ErrorCode.PRODUCT_NOT_FOUND));
     }
 
-    public Set<OptionResponse> getProductOptionsByIdOrThrow(Long id) {
-        return getProductByIdOrThrow(id).options();
-    }
-
     public ProductResponse createProduct(ProductRequest request) {
         Category category = categoryRepository.findByName(request.categoryName())
                 .orElseThrow(() -> new CustomException(ErrorCode.CATEGORY_NOT_FOUND));
@@ -60,31 +52,10 @@ public class ProductService {
         return ProductMapper.toResponseDto(product);
     }
 
-    @Transactional
-    public OptionResponse addOptionToProduct(Long id, OptionRequest request) {
-        Product product = productRepository.findById(id)
-                .orElseThrow(() -> new CustomException(ErrorCode.PRODUCT_NOT_FOUND));
-
-        return OptionMapper.toResponseDto(
-                optionRepository.save(OptionMapper.toEntity(request, product))
-        );
-    }
-
     public void deleteProductById(Long id) {
         productRepository.deleteById(id);
     }
 
-    @Transactional
-    public void deleteOptionFromProduct(Long id, OptionRequest request) {
-        Product product = productRepository.findById(id)
-                .orElseThrow(() -> new CustomException(ErrorCode.PRODUCT_NOT_FOUND));
-        if (product.getOptions()
-                   .size() == 1) {
-            throw new CustomException(ErrorCode.OPTION_REMOVE_FAILED);
-        }
-
-        optionRepository.deleteByProduct_IdAndName(id, request.name());
-    }
 
     public void deleteAllProducts() {
         productRepository.deleteAll();
